refactor: move RecalibrationPlan to its own leaf module
The mixin extraction left set_position/set_tilt_position importing
RecalibrationPlan function-locally, because the enum lived in cover_base
which imports the movement mixin at module load (a cycle). Relocating the
enum to a dependency-free leaf module (matching the position_reporting.py /
travel_calculator.py convention of one enum per owning module) lets both
cover_base and cover_movement import it at module level, removing the two
function-local imports and the cycle entirely. RecalibrationAxis stays in
cover_base (used only there).
